Strip extra newlines from settings page textareas
Morph adds a newline to the start of <textarea> form fields, which means
that each time you save the form with Mechanize it will add an extra
newline to the start of any environment variable that _isn't_ being
changed. To avoid this we're stripping all textareas before letting them
into the database.